tests/subscription-token-v-1-factory.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { BigInt, Address } from "@graphprotocol/graph-ts"
import { DeployFeeChange } from "../generated/schema"
import { DeployFeeChange as DeployFeeChangeEvent } from "../generated/SubscriptionTokenV1Factory/SubscriptionTokenV1Factory"
import { handleDeployFeeChange } from "../src/subscription-token-v-1-factory"
import { createDeployFeeChangeEvent } from "./subscription-token-v-1-factory-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let amount = BigInt.fromI32(234)
    let newDeployFeeChangeEvent = createDeployFeeChangeEvent(amount)
    handleDeployFeeChange(newDeployFeeChangeEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("DeployFeeChange created and stored", () => {
    assert.entityCount("DeployFeeChange", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "DeployFeeChange",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "amount",
      "234"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
